Renaming the preferences domain and data layer
