+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/imp/RetrieveTokenUseCaseImp.kt
@@ -1,6 +1,6 @@
 package me.androidbox.domain.authentication.usecase.imp
 
-import me.androidbox.domain.authentication.token_repository.PreferenceRepository
+import me.androidbox.domain.authentication.preference.PreferenceRepository
 import me.androidbox.domain.authentication.usecase.RetrieveTokenUseCase
 import javax.inject.Inject
