+++ domain/src/main/java/me/androidbox/domain/authentication/usecase/imp/SaveTokenUseCaseImp.kt
@@ -1,6 +1,6 @@
 package me.androidbox.domain.authentication.usecase.imp
 
-import me.androidbox.domain.authentication.token_repository.PreferenceRepository
+import me.androidbox.domain.authentication.preference.PreferenceRepository
 import me.androidbox.domain.authentication.usecase.SaveTokenUseCase
 import javax.inject.Inject
